logic for rating system is working

+++ app/src/main/java/com/example/mediaapp/models/RatingForDatabase.kt
@@ -2,13 +2,13 @@ package com.example.mediaapp.models
 
 data class RatingForDatabase(
     val movieID: Long,
-    val rating: Int
+    val rating: Long
 ) {
     companion object {
         fun fromMap(map: Map<String, Any?>): RatingForDatabase {
             return RatingForDatabase(
                 map["movieID"] as? Long ?: -1,
-                map["rating"] as? Int ?: -1,
+                map["rating"] as? Long ?: -1,
             )
         }
     }
